requirements.txt
selenium>=4.20.0
pyperclip>=1.8.2
pillow>=10.0.0
